+++ python/v4api_sdk/batch_ops_actions.py
@@ -46,7 +46,7 @@ from ntnx_prism_py_client.models.prism.v4.operations.BatchSpecPayloadMetadataPat
 from ntnx_prism_py_client.models.prism.v4.operations.ActionType import ActionType
 
 
-from tme import Utils
+from tme.utils import Utils
 
 
 def confirm_entity(api, client, entity_name: str) -> str:
