+++ python/v4api_sdk/batch_ops_create.py
@@ -31,7 +31,7 @@ from ntnx_prism_py_client.models.prism.v4.operations.BatchSpecPayload import (
 
 from ntnx_prism_py_client.models.prism.v4.operations.ActionType import ActionType
 
-from tme import Utils
+from tme.utils import Utils
 
 
 def main():
