+++ python/v4api_sdk/create_categories.py
@@ -21,7 +21,7 @@ from ntnx_prism_py_client import ApiClient as PrismClient
 from ntnx_prism_py_client.rest import ApiException as PrismException
 
 
-from tme import Utils
+from tme.utils import Utils
 from tme import Config
