+++ python/v4api_sdk/requirements.txt
@@ -1,10 +1,10 @@
 requests==2.31.0
 ruff==0.1.15
-ntnx_vmm_py_client==4.0.1b1
-ntnx_lifecycle_py_client==4.0.1b1
-ntnx_prism_py_client==4.0.1b1
-ntnx_clustermgmt_py_client==4.0.1b2
-ntnx_aiops_py_client==4.0.3a2
-ntnx_networking_py_client==4.0.1b1
-ntnx_microseg_py_client==4.0.1b1
-ntnx_iam_py_client==4.0.1b1
+ntnx_vmm_py_client==4.0.1
+ntnx_lifecycle_py_client==4.0.1
+ntnx_prism_py_client==4.0.1
+ntnx_clustermgmt_py_client==4.0.1
+ntnx_aiops_py_client==4.0.1
+ntnx_networking_py_client==4.0.1
+ntnx_microseg_py_client==4.0.1
+ntnx_iam_py_client==4.0.1
